Move consensus state to lib
